fixing bch
