--- NOTES.txt
@@ -1,49 +1,0 @@
-go get -u github.com/spf13/cobra
-
-litmusctl config view // view the config
-
-litmusctl config set-account --endpoint --username --password //set
-
-litmusctl config use-account <endpoint>
-
-litmuctl config unset-account <endpoint>
-
-litmusctl config get-accounts
-
-// password should hidden --done
-// schema for agent connect
-// project creation --donw
-// pass creds as secret
-
-litmusctl agent connect --name --project --installation-mode --description --platform --namespace --service-account --yes
-
-apiVersion: v1
-clusters:
-- cluster:
-    certificate-authority: /home/raj/.minikube/ca.crt
-    server: https://172.17.0.2:8443
-  name: minikube
-contexts:
-- context:
-    cluster: minikube
-    user: minikube
-  name: minikube
-current-context: minikube
-kind: Config
-preferences: {}
-users:
-- name: minikube
-  user:
-    client-certificate: /home/raj/.minikube/profiles/minikube/client.crt
-    client-key: /home/raj/.minikube/profiles/minikube/client.key
-
-apiVersion: v1
-kind: Config
-current-account: ""
-accounts:
-- account:
-    username: ""
-    password: ""
-    token: ""
-    expiry_date: ""
-  endpoint: ""
