+++ pkg/cmd/version/version.go
@@ -13,20 +13,21 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package get
+package version
 
 import (
+	"fmt"
+	"os"
+
 	"github.com/spf13/cobra"
 )
 
-// getCmd represents the get command
-var GetCmd = &cobra.Command{
-	Use:   "get",
+// versionCmd represents the version command
+var VersionCmd = &cobra.Command{
+	Use:   "version",
 	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+	Long: ``,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("Litmusctl version: ", os.Getenv("CLIVersion"))
+	},
 }
