--- tmp.yaml
@@ -1,10 +1,0 @@
-accounts:
-- users:
-  - expires_in: "1625508271"
-    token: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MjU1MDgyNzEsInJvbGUiOiJhZG1pbiIsInVpZCI6ImQ4MjA0Zjk1LTlhMWEtNDA3YS1hMWFhLTc5ZDdhZGQ5MGRmNSIsInVzZXJuYW1lIjoiYWRtaW4ifQ.mSvklX013Kejkkj2ELEfxhjT_jhLLb61WzW5lbU5RGvxD-LeSXT5-JCcZRf44zqDp7pO598O-zeZYC98UsQt0g
-    username: admin
-  endpoint: https://litmus.preview.chaosnative.com
-apiVersion: v1
-current-account: https://litmus.preview.chaosnative.com
-current-user: admin
-kind: Config
